KITTI_2DObject_Detection_YOLO3_Pytorch/KITTI-2d-object-detection/requirements.txt
certifi==2019.3.9
cffi==1.12.2
cloudpickle==0.8.1
cycler==0.10.0
cytoolz==0.9.0.1
dask==1.2.0
decorator==4.4.0
imageio==2.5.0
kiwisolver==1.0.1
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
networkx==2.3
numpy==1.15.0
olefile==0.46
Pillow==6.0.0
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.2.1
Shapely==1.6.4.post2
six==1.12.0
toolz==0.9.0
torch==1.0.1
torchvision==0.2.2
tornado==6.0.2
wincertstore==0.2
